+++ providers/AuthProvider.tsx
@@ -7,8 +7,11 @@ import Animated, {
   withTiming,
 } from "react-native-reanimated";
 
+import AccountDeletionScreen from "@/components/screens/AccountDeletionScreen";
+import MaintenanceScreen from "@/components/screens/MaintenanceScreen";
 import SplashScreen from "@/components/SplashScreen";
 import { supabase } from "@/lib/supabase";
+import { useAppStatusStore } from "@/stores/useAppStatus";
 import { useAuthProviderStore } from "@/stores/useAuthProviderStore";
 import { useAuthStore } from "@/stores/useAuthStore";
 import { useThemeStore } from "@/stores/useThemeStore";
@@ -16,6 +19,16 @@ import { useThemeStore } from "@/stores/useThemeStore";
 export function AuthProvider({ children }: { children: React.ReactNode }) {
   const { hydrate } = useThemeStore();
   const { showSplash, initializeApp } = useAuthProviderStore();
+  const {
+    maintenance,
+    maintenanceMessage,
+    accountDeletionRequested,
+    deletionMessage,
+    loading: appStatusLoading,
+    checkAppStatus,
+  } = useAppStatusStore();
+
+  const { session } = useAuthStore();
 
   // valores compartidos en Reanimated
   const splashOpacity = useSharedValue(1);
@@ -46,7 +59,7 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
           useAuthStore.getState().setSession(session);
           useAuthStore.getState().setToken(session.access_token);
         }
-      }
+      },
     );
     return () => {
       subscription.subscription.unsubscribe();
@@ -69,6 +82,26 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
     init();
   }, [isMounted]);
 
+  useEffect(() => {
+    const runChecks = async () => {
+      const userId = useAuthStore.getState().userId;
+      await checkAppStatus(userId);
+    };
+    if (!showSplash) runChecks();
+  }, [showSplash, session]);
+
+  // useEffect(() => {
+  //   if (!session) return;
+
+  //   checkAppStatus(session.user.id);
+  // }, [session]);
+
+  // useEffect(() => {
+  //   if (!session || showSplash) return;
+
+  //   checkAppStatus(session.user.id);
+  // }, [session, showSplash]);
+
   // animaciones splash → contenido
   useEffect(() => {
     if (!showSplash) {
@@ -82,6 +115,15 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
       });
     }
   }, [showSplash]);
+  if (appStatusLoading) return null;
+
+  if (maintenance) {
+    return <MaintenanceScreen message={maintenanceMessage} />;
+  }
+
+  if (accountDeletionRequested) {
+    return <AccountDeletionScreen message={deletionMessage} />;
+  }
 
   return (
     <View className="flex-1 bg-background dark:bg-dark-background">