+++ CHANGELOG.md
@@ -5,11 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.5] - 2025-MM-DD #UNRELEASED
+## [1.0.5] - 2026-MM-DD #UNRELEASED
 
 ### Added
 
--Creayte Order Summary Screen for create order
+- Creayte Order Summary Screen for create order
+- Add Maintenance screen
+- Add DeleteAccountScreen
+- Store in AuthProvider to check the maintenance status and delete Account status
 
 ### Changed